backend/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
openai-whisper==20231117
numpy==1.26.3
scipy==1.12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
